+++ spark/src/main/kotlin/com/adevinta/spark/res/AnnotatedStringResource.kt
@@ -315,88 +315,6 @@ private fun spannableStringToAnnotatedString(
     }
 }
 
-public sealed class SparkStringAnnotations {
-    public sealed class Color : SparkStringAnnotations() {
-        public object Primary : Color() {
-            override fun toString(): String = "primary"
-        }
-
-        public object Secondary : Color() {
-            override fun toString(): String = "secondary"
-        }
-
-        public object Success : Color() {
-            override fun toString(): String = "success"
-        }
-
-        public object Alert : Color() {
-            override fun toString(): String = "alert"
-        }
-
-        public object Error : Color() {
-            override fun toString(): String = "error"
-        }
-
-        public object Info : Color() {
-            override fun toString(): String = "info"
-        }
-
-        public object Neutral : Color() {
-            override fun toString(): String = "neutral"
-        }
-
-        override fun toString(): String = "color"
-
-        public companion object
-    }
-
-    public sealed class Typography : SparkStringAnnotations() {
-        public object Title1 : Typography() {
-            override fun toString(): String = "title1"
-        }
-
-        public object Title2 : Typography() {
-            override fun toString(): String = "title2"
-        }
-
-        public object Title3 : Typography() {
-            override fun toString(): String = "title3"
-        }
-
-        public object BodyImportant : Typography() {
-            override fun toString(): String = "bodyImportant"
-        }
-
-        public object Body : Typography() {
-            override fun toString(): String = "body"
-        }
-
-        public object SmallImportant : Typography() {
-            override fun toString(): String = "smallImportant"
-        }
-
-        public object Small : Typography() {
-            override fun toString(): String = "small"
-        }
-
-        public object ExtraSmallImportant : Typography() {
-            override fun toString(): String = "extraSmallImportant"
-        }
-
-        public object ExtraSmall : Typography() {
-            override fun toString(): String = "extraSmall"
-        }
-
-        public object Button : Typography() {
-            override fun toString(): String = "button"
-        }
-
-        public companion object {
-            public const val keyName: String = "typography"
-        }
-    }
-}
-
 @Preview
 @Preview(
     locale = "fr-rFR",
